LangChain/orchestrator-chat-multisessions-inmemory/requirements.txt
Flask==2.3.1
jsonify==0.5
langchain==0.0.138
load-dotenv==0.1.0
openai==0.27.4
python-dotenv==1.0.0
requests==2.28.2
tiktoken==0.3.3
waitress==2.1.2